renamed files to match components
